# api/docs/swagger.yaml
definitions:
  models.Animal:
    properties:
      animal_type_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      disease:
        type: string
      gender:
        type: string
      id:
        type: string
      last_fed_time:
        type: string
      last_watered_time:
        type: string
      name:
        type: string
      updated_at:
        type: string
      weight:
        type: integer
    type: object
...
  models.CreateFeeding:
    properties:
      animal_id:
        type: string
      male_id:
        type: string
    type: object
  models.Feeding:
    properties:
      animal_id:
        type: string
      feeding_time:
        type: string
      id:
        type: string
      male_id:
        type: string
    type: object
  models.FoodWarehouse:
    properties:
      animal_id:
        type: string
      animal_type_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      group_feeding:
        type: boolean
      id:
        type: string
      name:
        type: string
      quantity:
        type: integer
      quantity_type:
        type: string
      updated_at:
        type: string
    type: object
  models.FoodWarehouseCreate:
    properties:
      animal_id:
        type: string
      animal_type_id:
        type: string
      group_feeding:
        type: boolean
      name:
        type: string
      quantity:
        type: integer
      quantity_type:
        type: string
    type: object
...
  models.ListFeeding:
    properties:
      count:
        type: integer
      feedings:
        items:
          $ref: '#/definitions/models.Feeding'
        type: array
    type: object
  models.ListFoodWarehouse:
    properties:
      count:
        type: integer
      food_warehouses:
        items:
          $ref: '#/definitions/models.FoodWarehouse'
        type: array
    type: object
...
  models.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  models.StatusRes:
    properties:
      status:
        type: boolean
    type: object
...
  models.UpdateFoodWarehouseReq:
    properties:
      animal_id:
        type: string
      animal_type:
        type: string
      group_feeding:
        type: boolean
      id:
        type: string
      name:
        type: string
      quantity:
        type: integer
      quantity_type:
        type: string
    type: object
...
host: localhost:9050
info:
  contact: {}
  title: CRM Farmish
  version: "1.7"
paths:
...
  /v1/feeding:
    get:
      consumes:
      - application/json
      description: ListFeedingBy - API for listing feedings
      parameters:
      - description: limit
        in: query
        name: limit
        type: string
      - description: page
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListFeeding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List Feeding
      tags:
      - Feeding
    post:
      consumes:
      - application/json
      description: CreateFeeding - Api for crete Feeding
      parameters:
      - description: AnimalTypeCreate
        in: body
        name: AnimalTypeCreate
        required: true
        schema:
          $ref: '#/definitions/models.CreateFeeding'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Animal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create Feeding
      tags:
      - Feeding
  /v1/feeding/animal-id:
    get:
      consumes:
      - application/json
      description: ListFeedingByAnimalID - Api for List Feeding By AnimalID
      parameters:
      - in: query
        name: animal_id
        type: string
      - in: query
        name: delete_status
        type: boolean
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListFeeding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List Feeding By AnimalID
      tags:
      - Feeding
  /v1/feeding/get:
    get:
      consumes:
      - application/json
      description: GetFeeding - Api for Get Feeding
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Feeding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get Feeding
      tags:
      - Feeding
  /v1/food:
    delete:
      consumes:
      - application/json
      description: DeleteFoodWarehouse - Api for Delete FoodWarehouse
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete FoodWarehouse
      tags:
      - Food Warehouse
    get:
      consumes:
      - application/json
      description: ListFoodWarehouse - Api for List FoodWarehouse
      parameters:
      - in: query
        name: delete_status
        type: boolean
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListFoodWarehouse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List Food Warehouse
      tags:
      - Food Warehouse
    post:
      consumes:
      - application/json
      description: CreateFoodWarehouse - Api for crete FoodWarehouse
      parameters:
      - description: FoodWarehouseCreate
        in: body
        name: FoodWarehouseCreate
        required: true
        schema:
          $ref: '#/definitions/models.FoodWarehouseCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FoodWarehouse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create Food Warehouse
      tags:
      - Food Warehouse
    put:
      consumes:
      - application/json
      description: UpdateFoodWarehouse - Api for Update FoodWarehouse
      parameters:
      - description: UpdateAnimalTypeReq
        in: body
        name: UpdateAnimalTypeReq
        required: true
        schema:
          $ref: '#/definitions/models.UpdateFoodWarehouseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FoodWarehouse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update FoodWarehouse
      tags:
      - Food Warehouse
  /v1/food/get:
    get:
      consumes:
      - application/json
      description: GetFoodWarehouse - Api for Get FoodWarehouse
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FoodWarehouse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get Food Warehouse
      tags:
      - Food Warehouse
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"